Skip NullChecks in LSE when replacing loads
Due to how LSE works, if a set instruction sets a value which is
a NullCheck we record the NullCheck instruction. We use that for
verification purposes within LSE too. However, when performing
the load replacement we don't care about that NullCheck so we can
skip it and fetch what it is NullChecking (probably a Load or Phi).

Bug: 243136201
Test: dex2oat compiling locally the app in the bug
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Change-Id: I10a363af6109ee6f73d28d0fa4305cc8a57431b8
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: tejasvp25 <tejasvp25@gmail.com>
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>